requirements.txt
pyyaml
addict
prompt_toolkit
openai>=0.27.0